perf(sync): index the new reconcile predicates and batch the pull path
The category-settings sync added two reconcile fetches that couldn't be
served by an index: LiveStream `isFavorite || isHidden` (isHidden was only
indexed behind categoryId, so SQLite's OR optimization couldn't use it and
every reconcile scanned the whole channel table) and Category
`isHidden || customOrder != nil` (only isRestricted was indexed). Add
standalone isHidden/customOrder indexes so both fetches seek the handful
of customized rows.

Also replace the pull path's one-fetch-per-id model resolution with one
chunked IN fetch per kind: a fresh device pulling a whole profile's
states, and every profile switch (importProfileState), previously issued
one single-row fetch per record — bulk-hidden channels/categories made
that set much larger. A batch miss now counts as pending directly instead
of re-fetching per id.

Content mutation helpers move to CloudSyncEngine+Content.swift to keep
CloudSyncEngine.swift within the 600-line cap.
