+++ Lume/Services/Sync/CloudSync/CloudSyncEngine+Profiles.swift
@@ -163,8 +163,18 @@ private extension CloudSyncEngine {
     }
 
     func importProfileState(_ profileID: UUID) throws {
-        for (id, mirror) in try fetchMirrors(forProfile: profileID) {
-            _ = try applyContentToLocal(Self.values(from: mirror), id: id, kind: mirror.kind, loaded: nil)
+        let mirrors = try fetchMirrors(forProfile: profileID)
+        var idsByKind: [SyncedContentKind: [String]] = [:]
+        for (id, mirror) in mirrors {
+            idsByKind[mirror.kind, default: []].append(id)
+        }
+        // One batched fetch per kind instead of one single-row fetch per mirror.
+        // A mirror whose catalog item hasn't synced here yet simply doesn't
+        // project — same outcome as the old per-id miss, without the fetch.
+        let loaded = try fetchCatalogModels(byKind: idsByKind)
+        for (id, mirror) in mirrors {
+            guard let model = loaded[id] else { continue }
+            _ = try applyContentToLocal(Self.values(from: mirror), id: id, kind: mirror.kind, loaded: model)
         }
     }
