+++ Lume/Services/Sync/CloudSync/CloudSyncEngine.swift
@@ -186,6 +186,13 @@ actor CloudSyncEngine {
         var ids = Set(mirrors.keys).union(localValues.keys)
         ids.formUnion(shadow.contentShadowIDs())
 
+        // Pull verdicts whose catalog model isn't already in `localValues` are
+        // deferred and resolved with one batched fetch per kind below — a fresh
+        // device pulling a whole profile's states would otherwise issue one
+        // single-row fetch per id.
+        var deferred: [(id: String, verdict: MergeVerdict<ContentStateValues>)] = []
+        var deferredIDs: [SyncedContentKind: [String]] = [:]
+
         for id in ids {
             // Garbage-collect state whose owning playlist no longer exists on
             // either side (deleted however). Clear the cloud record, reset any
@@ -203,6 +210,11 @@ actor CloudSyncEngine {
                 shadow: shadow.contentShadow(id),
                 mergeConflict: ContentStateValues.mergeConflict
             )
+            if verdict.writesLocal, localValues[id] == nil, let kind = mirrors[id]?.kind {
+                deferred.append((id, verdict))
+                deferredIDs[kind, default: []].append(id)
+                continue
+            }
             try applyContentVerdict(
                 verdict,
                 id: id,
@@ -210,6 +222,18 @@ actor CloudSyncEngine {
                 loaded: localValues[id]?.model,
                 into: &result
             )
+        }
+
+        let loaded = try fetchCatalogModels(byKind: deferredIDs)
+        for (id, verdict) in deferred {
+            // The batch fetch is authoritative: a miss means the catalog item
+            // hasn't synced to this device yet — count it pending (shadow
+            // untouched, same as the old per-id miss) without re-fetching.
+            guard let model = loaded[id] else {
+                result.contentPending += 1
+                continue
+            }
+            try applyContentVerdict(verdict, id: id, mirror: mirrors[id], loaded: model, into: &result)
         }
     }
 
@@ -251,6 +275,17 @@ actor CloudSyncEngine {
 }
 
 // MARK: - Verdict application
+
+private extension MergeVerdict {
+    /// True for verdicts that write to the local catalog and therefore need the
+    /// catalog model resolved before they can be applied.
+    var writesLocal: Bool {
+        switch self {
+        case .pullToLocal, .writeBoth: true
+        case .noChange, .pushToCloud: false
+        }
+    }
+}
 
 private extension CloudSyncEngine {
     func applyPlaylistVerdict(
@@ -446,154 +481,5 @@ private extension CloudSyncEngine {
     }
 }
 
-// MARK: - Content mutations
-
-/// Not `private`: profile operations in `CloudSyncEngine+Profiles.swift`
-/// reuse these helpers (fetch / reset / apply / value extraction).
-extension CloudSyncEngine {
-    static func kind(of model: (any PersistentModel)?) -> SyncedContentKind? {
-        switch model {
-        case is Movie: .movie
-        case is Series: .series
-        case is Episode: .episode
-        case is LiveStream: .live
-        case is Category: .category
-        default: nil
-        }
-    }
-
-    func applyContentToCloud(_ value: ContentStateValues?, id: String, kind: SyncedContentKind?, mirror: UserContentState?) {
-        guard let value, !value.isEmpty else {
-            if let mirror { cloudContext.delete(mirror) }
-            return
-        }
-        let kind = kind ?? mirror?.kind ?? .movie
-        if let mirror {
-            mirror.profileID = activeProfileID // heals a legacy nil record on first touch
-            mirror.kindRaw = kind.rawValue
-            mirror.watchProgress = value.watchProgress
-            mirror.isWatched = value.isWatched
-            mirror.lastWatchedDate = value.lastWatchedDate
-            mirror.isFavorite = value.isFavorite
-            mirror.addedToWatchlistDate = value.addedToWatchlistDate
-            mirror.favoriteOrder = value.favoriteOrder
-            mirror.recommendationVoteRaw = value.recommendationVoteRaw
-            mirror.isHidden = value.isHidden
-            mirror.customOrder = value.customOrder
-            mirror.updatedAt = Date()
-        } else {
-            cloudContext.insert(UserContentState(
-                contentId: id,
-                kind: kind,
-                profileID: activeProfileID,
-                watchProgress: value.watchProgress,
-                isWatched: value.isWatched,
-                lastWatchedDate: value.lastWatchedDate,
-                isFavorite: value.isFavorite,
-                addedToWatchlistDate: value.addedToWatchlistDate,
-                favoriteOrder: value.favoriteOrder,
-                recommendationVoteRaw: value.recommendationVoteRaw,
-                isHidden: value.isHidden,
-                customOrder: value.customOrder
-            ))
-        }
-    }
-
-    /// Applies a cloud value to the matching local catalog item. Returns false
-    /// (without touching the shadow) when the catalog item hasn't synced to this
-    /// device yet, so the change stays pending for a later pass.
-    func applyContentToLocal(_ value: ContentStateValues?, id: String, kind: SyncedContentKind?, loaded: (any PersistentModel)?) throws -> Bool {
-        guard let kind else { return true } // nothing to apply (shadow-only id)
-        let values = value ?? ContentStateValues(watchProgress: 0, isWatched: false, lastWatchedDate: nil, isFavorite: false, addedToWatchlistDate: nil, favoriteOrder: nil)
-
-        // Each helper returns false when its catalog item hasn't synced yet, so
-        // the change stays pending for a later pass.
-        switch kind {
-        case .movie: return try applyMovieToLocal(values, id: id, loaded: loaded)
-        case .series: return try applySeriesToLocal(values, id: id, loaded: loaded)
-        case .episode: return try applyEpisodeToLocal(values, id: id, loaded: loaded)
-        case .live: return try applyLiveToLocal(values, id: id, loaded: loaded)
-        case .category: return try applyCategoryToLocal(values, id: id, loaded: loaded)
-        }
-    }
-
-    private func applyMovieToLocal(_ values: ContentStateValues, id: String, loaded: (any PersistentModel)?) throws -> Bool {
-        guard let movie = try (loaded as? Movie) ?? fetchMovie(id) else { return false }
-        movie.watchProgress = values.watchProgress
-        movie.isWatched = values.isWatched
-        movie.lastWatchedDate = values.lastWatchedDate
-        movie.isFavorite = values.isFavorite
-        movie.addedToWatchlistDate = values.addedToWatchlistDate
-        movie.recommendationVoteRaw = values.recommendationVoteRaw
-        return true
-    }
-
-    private func applySeriesToLocal(_ values: ContentStateValues, id: String, loaded: (any PersistentModel)?) throws -> Bool {
-        guard let series = try (loaded as? Series) ?? fetchSeries(id) else { return false }
-        series.isFavorite = values.isFavorite
-        series.addedToWatchlistDate = values.addedToWatchlistDate
-        series.lastWatchedDate = values.lastWatchedDate
-        series.recommendationVoteRaw = values.recommendationVoteRaw
-        return true
-    }
-
-    private func applyEpisodeToLocal(_ values: ContentStateValues, id: String, loaded: (any PersistentModel)?) throws -> Bool {
-        guard let episode = try (loaded as? Episode) ?? fetchEpisode(id) else { return false }
-        episode.watchProgress = values.watchProgress
-        episode.isWatched = values.isWatched
-        episode.lastWatchedDate = values.lastWatchedDate
-        return true
-    }
-
-    private func applyLiveToLocal(_ values: ContentStateValues, id: String, loaded: (any PersistentModel)?) throws -> Bool {
-        guard let stream = try (loaded as? LiveStream) ?? fetchLiveStream(id) else { return false }
-        stream.isFavorite = values.isFavorite
-        stream.favoriteOrder = values.favoriteOrder
-        stream.isHidden = values.isHidden
-        // `customOrder` (per-category channel order) is intentionally
-        // device-local — not pulled from the mirror.
-        return true
-    }
-
-    private func applyCategoryToLocal(_ values: ContentStateValues, id: String, loaded: (any PersistentModel)?) throws -> Bool {
-        guard let category = try (loaded as? Category) ?? fetchCategory(id) else { return false }
-        category.isHidden = values.isHidden
-        category.customOrder = values.customOrder
-        return true
-    }
-
-    /// Resets an orphaned local item's user state to defaults so it stops
-    /// regenerating cloud records after its playlist was deleted.
-    func resetLocalContent(_ entry: LocalContentEntry) {
-        switch entry.model {
-        case let movie as Movie:
-            movie.watchProgress = 0
-            movie.isWatched = false
-            movie.lastWatchedDate = nil
-            movie.isFavorite = false
-            movie.addedToWatchlistDate = nil
-            movie.recommendationVoteRaw = 0
-        case let series as Series:
-            series.isFavorite = false
-            series.addedToWatchlistDate = nil
-            series.lastWatchedDate = nil
-            series.recommendationVoteRaw = 0
-        case let episode as Episode:
-            episode.watchProgress = 0
-            episode.isWatched = false
-            episode.lastWatchedDate = nil
-        case let stream as LiveStream:
-            stream.isFavorite = false
-            stream.favoriteOrder = nil
-            stream.isHidden = false
-        // `customOrder` is device-local (not projected per profile) — leave it.
-        case let category as Category:
-            // `isRestricted` is a device-global parental control, not synced
-            // per-profile state — leave it untouched.
-            category.isHidden = false
-            category.customOrder = nil
-        default:
-            break
-        }
-    }
-}
+// The per-content mutation helpers (`applyContentToCloud` / `applyContentToLocal`
+// / `resetLocalContent`) live in CloudSyncEngine+Content.swift.
